+++ src/components/Home/Home.js
@@ -1,47 +1,68 @@
 import React from "react"
-import { Snip1104Blue, Snip1104Yelow } from "./HomeStyle"
-import { AnimatedContainer, ContainerHome } from "./HomeStyle"
-import hairBlueBg from "../../images/comCabeloBlue.jpg"
-import hairYelowBg from "../../images/comCabeloYelow.jpg"
+import { ContainerHome } from "./HomeStyle"
 import { Carousel } from "../Carousel/Carousel"
 import { Testimonial } from "../Testimonial/Testimonial"
 import { Media } from "../Media/Media"
 import { Title } from "../Title/Title"
+import { ChooseNinja } from "../ChooseNinja/ChooseNinja"
+import Cart from "../Cart/Cart"
 
 
+export default class Home extends React.Component {
+    state = {
+        currentPage: ''
+    }
 
-export const Home = () => {
-    return (
-        <>
-            <Title />
+    changePage = (currentPage) => {
+        this.setState({ currentPage: currentPage })
+    }
 
+    render() {
+
+        const renderCurrentPage = () => {
+            switch (this.state.currentPage) {
+                case 'home':
+                    return ""
+                case 'cart':
+                    return <Cart />
+            }
+        }
+
+        return (
             <ContainerHome>
-                <AnimatedContainer>
-                    <Snip1104Blue>
-                        <img src={hairBlueBg} alt="ninja face" />
-                        <figcaption>
-                            <h2>Seja um<span> Ninja</span></h2>
-                        </figcaption>
-                        {/* <a href="/"></a> */}
-                       
-                    </Snip1104Blue>
 
-                    <Snip1104Yelow>
-                        <img src={hairYelowBg} alt="ninja face" />
-                        <figcaption>
-                            <h2>Contrate um<span> Ninja</span></h2>
-                        </figcaption>
-                        {/* <a href="/"></a> */}
-                    </Snip1104Yelow>
-                </AnimatedContainer>
+                <Title />
+
+                <ChooseNinja
+                    changePage={this.changePage} />
+                {renderCurrentPage()}
+
+                <Carousel />
+
+                <Testimonial />
+
+                <Media />
+
             </ContainerHome>
-
-            <Carousel />
-
-            <Testimonial />
-
-            <Media />
-
-        </>
-    )
+        )
+    }
 }
+
+// export const Home = () => {
+//     return (
+//         <ContainerHome>
+
+//             <Title />
+
+//             <ChooseNinja />
+
+//             <Carousel />
+
+//             <Testimonial />
+
+//             <Media />
+
+//         </ContainerHome>
+//     )
+// }
+
