+++ src/components/Media/Mediatyle.js
@@ -3,11 +3,25 @@ import styled from "styled-components";
 
 export const MediaContainer = styled.div`
     background-color: #111835;
-    height: 30vh;    
-    color: #f3cb11;
-    padding: 8px;
+    height: 30vh;      
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+`
 
+export const TituloMedia = styled.div`
+    color: #f3cb11;
+    padding: 16px; 
+    font-size: 2rem;
+`
+
+export const ImagensMediaContainer = styled.div`
     display: flex;
     align-items: center;
-    align-content: center;
+    justify-content: center;
+        
+    img {       
+        width: 200px        
+    }
 `
