+++ src/components/Testimonial/TestimonialStyle.js
@@ -3,8 +3,7 @@ import styled from "styled-components";
 
 export const TestimonialContainer = styled.div`
     background-color: #f8d613;
-    height: 80vh;
-    width: 100vw;
+    height: 80vh;    
     color: #0448b7;
     padding: 8px;
 `
